TASK: extract LinkOption into editors package
