+++ packages/neos-ui-editors/src/Editors/Reference/index.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import SelectBox from '@neos-project/react-ui-components/src/SelectBox/';
 import dataLoader from './referenceDataLoader';
 import createNew from './createNew';
-import NodeOption from '@neos-project/neos-ui-ckeditor-bindings/src/EditorToolbar/NodeOption';
+import NodeOption from '../../Library/NodeOption';
 import {neos} from '@neos-project/neos-ui-decorators';
 
 @neos(globalRegistry => ({
