+++ packages/neos-ui-editors/src/Library/AssetOption.js
