+++ packages/neos-ui-editors/src/Library/NodeOption.js
